requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
groq==0.4.1
flask-cors==4.0.0
gunicorn==21.2.0
